requirements.txt
fastapi
uvicorn
websockets
pytesseract
pillow
faster-whisper
python-multipart
numpy
ollama
soundfile
torch
silero-vad